fix: address Copilot re-review on PR #1099 (safety-offset, IsLoading, int loops)
- ImageWorldMapCore.ImportBorder: enforce U.isSafetyOffset(borderRecordAddr) so a
  caller cannot pass a ROM-header/danger-zone address (e.g. 0x100) and corrupt the
  header via the P0/P4 repoint. Added a theory test (0 / 0x100 / 0x1FF → reject,
  ZERO mutation) alongside the existing past-EOF case.
- WorldMapImageView.DoBorderImport: wrap the post-import LoadBorderEntry reload in
  an IsLoading scope and MarkClean AFTER it, so the SetField writes don't flip
  IsDirty back true after a committed import.
- ImageUtilBorderAPCore.BuildApData: switch the two frame OAM-copy loops from uint
  to int indices (clearer byte[] indexing; the split counts are small int-range
  totals) — addresses the reviewer's index-type concern.

Claude Code v2.1.168 · model claude-opus-4-8[1m]

+++ FEBuilderGBA.Core/ImageUtilBorderAPCore.cs
@@ -526,15 +526,16 @@ namespace FEBuilderGBA
             AppendU16(newOam, 0);
             AppendU16(newOam, 0);
 
-            // frame_0.
+            // frame_0. (int loop indices — apOam is byte[]; the split counts are
+            // small OAM byte totals well within int range.)
             uint addrFrame0 = (uint)newOam.Count;
             AppendU16(newOam, apOamSplit[0] / 6); // oam entry count
-            for (uint i = 0; i < apOamSplit[0]; i++) newOam.Add(apOam[i]);
+            for (int i = 0; i < (int)apOamSplit[0]; i++) newOam.Add(apOam[i]);
 
             // frame_1.
             uint addrFrame1 = (uint)newOam.Count;
             AppendU16(newOam, (apOamSplit[1] - apOamSplit[0]) / 6);
-            for (uint i = apOamSplit[0]; i < apOamSplit[1]; i++) newOam.Add(apOam[i]);
+            for (int i = (int)apOamSplit[0]; i < (int)apOamSplit[1]; i++) newOam.Add(apOam[i]);
 
             // anim_0.
             uint addrAnim0 = (uint)newOam.Count;
